fix eslint warnings, move some components which are only used once out of separate files
