+++ frontend/src/components/ReviewDetails.tsx
@@ -1,11 +1,9 @@
 import {Review} from "../lib/Review";
-import React, {useEffect, useState} from "react";
+import React, {useEffect, useState, useCallback} from "react";
 import {CommentRequest, CommentService, ReviewComment} from "../lib/ReviewComment";
-import {CommentList} from "./CommentList";
 import {ReviewCardWithoutLink} from "./ReviewCards";
 import "./ReviewDetails.css"
 import {useParams} from "react-router-dom";
-import useUser from "../hooks/useUser";
 
 interface ReviewDetailsProps {
     review: Review
@@ -15,10 +13,11 @@ export function ReviewDetails({review}: ReviewDetailsProps) {
     const {courseId, authorUsername} = useParams();
     const [comments, setComments] = useState<ReviewComment[]>([]);
     const [newComment, setNewComment] = useState<string>("");
+    const memorizedReloadComments = useCallback(reloadComments, [review])
 
     useEffect(() => {
-        reloadComments()
-    }, [review, courseId, authorUsername]);
+        memorizedReloadComments()
+    }, [review, courseId, authorUsername, memorizedReloadComments]);
 
     function reloadComments() {
         CommentService.fetchCommentsByReview(review)
@@ -29,7 +28,7 @@ export function ReviewDetails({review}: ReviewDetailsProps) {
     }
 
     function handleCommentSubmit() {
-        if (newComment == "") return; //todo: inform user comment can't be blank
+        if (newComment === "") return; //todo: inform user comment can't be blank
         const request: CommentRequest = {
             text: newComment
         }
@@ -54,4 +53,28 @@ export function ReviewDetails({review}: ReviewDetailsProps) {
             <button onClick={handleCommentSubmit}>Dodaj komentarz</button>
         </div>
     </div>
+}
+
+interface CommentListProps {
+    comments: ReviewComment[]
+}
+
+export function CommentList({comments}: CommentListProps) {
+    return <ul>
+        {comments //todo: sort
+            .map(c => <li
+                key={c.id}><CommentCard review={c}/>
+            </li>)}
+    </ul>
+}
+
+interface CommentCardProps {
+    review: ReviewComment;
+}
+
+function CommentCard({review}: CommentCardProps) {
+    return <>
+        <div>{review.authorUsername}</div>
+        <div>{review.text}</div>
+    </>
 }
